refactor: flatten auth folders - remove the additional folder layer for each auth type
refactor: move auth components to auth folder
