+++ src/components/auth/ApiKeyAuth/ApiKeyAuthFlow.tsx
@@ -1,12 +1,13 @@
 import { useEffect, useState } from 'react';
 import { GenerateConnectionRequest, ProviderInfo } from '@generated/api/src';
 
-import { LandingContent } from 'components/BasicAuth/BasicAuthFlow/LandingContent';
 import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 import { api, Connection } from 'services/api';
 
-type BasicAuthFlowProps = {
+import { LandingContent } from './LandingContent';
+
+type ApiKeyAuthFlowProps = {
   provider: string;
   providerInfo: ProviderInfo;
   consumerRef: string;
@@ -18,32 +19,24 @@ type BasicAuthFlowProps = {
   setSelectedConnection: (connection: Connection | null) => void;
 };
 
-type BasicCreds = {
-  user: string;
-  pass: string;
-};
-
-export function BasicAuthFlow({
-  provider, providerInfo, consumerRef, consumerName, groupRef, groupName,
-  children, selectedConnection, setSelectedConnection,
-}: BasicAuthFlowProps) {
+export function ApiKeyAuthFlow({
+  provider, providerInfo, consumerRef, consumerName, groupRef, groupName, children,
+  selectedConnection, setSelectedConnection,
+}: ApiKeyAuthFlowProps) {
   const project = useProject();
   const [nextStep, setNextStep] = useState<boolean>(false);
-  const [creds, setCreds] = useState<BasicCreds | null>(null);
+  const [providerApiKey, setProviderApiKey] = useState<string | null>(null);
   const apiKey = useApiKey();
 
   useEffect(() => {
-    if (provider && api && nextStep && creds != null) {
+    if (provider && api && nextStep && providerApiKey != null) {
       const req: GenerateConnectionRequest = {
         groupName,
         groupRef,
         consumerName,
         consumerRef,
         provider,
-        basicAuth: {
-          username: creds.user,
-          password: creds.pass,
-        },
+        apiKey: providerApiKey,
       };
 
       api().connectionApi.generateConnection({ projectIdOrName: project.projectId, generateConnectionParams: req }, {
@@ -54,23 +47,16 @@ export function BasicAuthFlow({
         console.error('Error loading provider info: ', err);
       });
     }
-  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName,
-    groupRef, project, creds, setSelectedConnection]);
+  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName, groupRef,
+    project, setSelectedConnection, providerApiKey]);
 
-  const onNext = (user: string, pass: string) => {
-    setCreds({ user, pass });
+  const onNext = (value: string) => {
+    setProviderApiKey(value);
     setNextStep(true);
   };
 
   if (selectedConnection === null) {
-    return (
-      <LandingContent
-        provider={provider}
-        providerInfo={providerInfo}
-        handleSubmit={onNext}
-        error={null}
-      />
-    );
+    return <LandingContent provider={provider} providerInfo={providerInfo} handleSubmit={onNext} error={null} />;
   }
 
   return children;
