+++ src/components/auth/AuthErrorAlert/AuthErrorAlert.tsx
