+++ src/components/auth/BasicAuth/BasicAuthFlow.tsx
@@ -1,39 +1,50 @@
 import { useEffect, useState } from 'react';
-import { GenerateConnectionRequest } from '@generated/api/src';
+import { GenerateConnectionRequest, ProviderInfo } from '@generated/api/src';
 
-import { LandingContent } from 'components/NoAuth/NoAuthFlow/LandingContent';
 import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 import { api, Connection } from 'services/api';
 
-type NoAuthFlowProps = {
+import { LandingContent } from './LandingContent';
+
+type BasicAuthFlowProps = {
   provider: string;
+  providerInfo: ProviderInfo;
   consumerRef: string;
   consumerName?: string;
   groupRef: string;
   groupName?: string;
-  providerName?: string;
   children: JSX.Element,
   selectedConnection: Connection | null;
   setSelectedConnection: (connection: Connection | null) => void;
 };
 
-export function NoAuthFlow({
-  provider, consumerRef, consumerName, groupRef, groupName,
-  children, selectedConnection, setSelectedConnection, providerName,
-}: NoAuthFlowProps) {
+type BasicCreds = {
+  user: string;
+  pass: string;
+};
+
+export function BasicAuthFlow({
+  provider, providerInfo, consumerRef, consumerName, groupRef, groupName,
+  children, selectedConnection, setSelectedConnection,
+}: BasicAuthFlowProps) {
   const project = useProject();
-  const apiKey = useApiKey();
   const [nextStep, setNextStep] = useState<boolean>(false);
+  const [creds, setCreds] = useState<BasicCreds | null>(null);
+  const apiKey = useApiKey();
 
   useEffect(() => {
-    if (provider && api && nextStep) {
+    if (provider && api && nextStep && creds != null) {
       const req: GenerateConnectionRequest = {
         groupName,
         groupRef,
         consumerName,
         consumerRef,
         provider,
+        basicAuth: {
+          username: creds.user,
+          password: creds.pass,
+        },
       };
 
       api().connectionApi.generateConnection({ projectIdOrName: project.projectId, generateConnectionParams: req }, {
@@ -44,14 +55,23 @@ export function NoAuthFlow({
         console.error('Error loading provider info: ', err);
       });
     }
-  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName, groupRef, setSelectedConnection, project]);
+  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName,
+    groupRef, project, creds, setSelectedConnection]);
 
-  const onNext = () => {
+  const onNext = (user: string, pass: string) => {
+    setCreds({ user, pass });
     setNextStep(true);
   };
 
   if (selectedConnection === null) {
-    return <LandingContent handleSubmit={onNext} error={null} providerName={providerName} />;
+    return (
+      <LandingContent
+        provider={provider}
+        providerInfo={providerInfo}
+        handleSubmit={onNext}
+        error={null}
+      />
+    );
   }
 
   return children;
