+++ src/components/auth/NoAuth/NoAuthFlow.tsx
@@ -1,41 +1,40 @@
 import { useEffect, useState } from 'react';
-import { GenerateConnectionRequest, ProviderInfo } from '@generated/api/src';
+import { GenerateConnectionRequest } from '@generated/api/src';
 
-import { LandingContent } from 'components/ApiKeyAuth/ApiKeyAuthFlow/LandingContent';
 import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
 import { api, Connection } from 'services/api';
 
-type ApiKeyAuthFlowProps = {
+import { LandingContent } from './LandingContent';
+
+type NoAuthFlowProps = {
   provider: string;
-  providerInfo: ProviderInfo;
   consumerRef: string;
   consumerName?: string;
   groupRef: string;
   groupName?: string;
+  providerName?: string;
   children: JSX.Element,
   selectedConnection: Connection | null;
   setSelectedConnection: (connection: Connection | null) => void;
 };
 
-export function ApiKeyAuthFlow({
-  provider, providerInfo, consumerRef, consumerName, groupRef, groupName, children,
-  selectedConnection, setSelectedConnection,
-}: ApiKeyAuthFlowProps) {
+export function NoAuthFlow({
+  provider, consumerRef, consumerName, groupRef, groupName,
+  children, selectedConnection, setSelectedConnection, providerName,
+}: NoAuthFlowProps) {
   const project = useProject();
-  const [nextStep, setNextStep] = useState<boolean>(false);
-  const [providerApiKey, setProviderApiKey] = useState<string | null>(null);
   const apiKey = useApiKey();
+  const [nextStep, setNextStep] = useState<boolean>(false);
 
   useEffect(() => {
-    if (provider && api && nextStep && providerApiKey != null) {
+    if (provider && api && nextStep) {
       const req: GenerateConnectionRequest = {
         groupName,
         groupRef,
         consumerName,
         consumerRef,
         provider,
-        apiKey: providerApiKey,
       };
 
       api().connectionApi.generateConnection({ projectIdOrName: project.projectId, generateConnectionParams: req }, {
@@ -46,16 +45,14 @@ export function ApiKeyAuthFlow({
         console.error('Error loading provider info: ', err);
       });
     }
-  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName, groupRef,
-    project, setSelectedConnection, providerApiKey]);
+  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName, groupRef, setSelectedConnection, project]);
 
-  const onNext = (value: string) => {
-    setProviderApiKey(value);
+  const onNext = () => {
     setNextStep(true);
   };
 
   if (selectedConnection === null) {
-    return <LandingContent provider={provider} providerInfo={providerInfo} handleSubmit={onNext} error={null} />;
+    return <LandingContent handleSubmit={onNext} error={null} providerName={providerName} />;
   }
 
   return children;
