+++ src/components/auth/Oauth/OAuthWindow/windowHelpers.tsx
